% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_date_calmonth_calyear.R
\name{date_to_calmonth_c}
\alias{date_to_calmonth_c}
\title{Date -> calmonth (character)}
\usage{
date_to_calmonth_c(x = lubridate::today())

date_to_calmonth_c(x = lubridate::today())
}
\arguments{
\item{x}{a Date object or string, in the form of 'yyyy-mm-dd'}
}
\value{
calmonth ("XX")

calyearmonth ("YYYYMXX")
}
\description{
Date -> calmonth (character)

Date -> calyearmonth (character)
}
\examples{
date_to_calmonth_c('2021-08-11')
date_to_calmonth_c(lubridate::today())
date_to_calyearmonth_c('2021-08-11')
date_to_calyearmonth_c(lubridate::today())
}
