% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_date_calmonth_calyear.R
\name{date_to_calyear_n}
\alias{date_to_calyear_n}
\title{Date -> calyear (numeric)}
\usage{
date_to_calyear_n(x = lubridate::today())
}
\arguments{
\item{x}{a Date object or string, in the form of 'yyyy-mm-dd'}
}
\value{
ISO year in character
}
\description{
Date -> calyear (numeric)
}
\examples{
date_to_calyear_n('2021-08-11')
date_to_calyear_n(lubridate::today())
}
