% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_date_isoweek_isoyear.R
\name{date_to_isoquarter_n}
\alias{date_to_isoquarter_n}
\alias{date_to_isoquarter_n.default}
\alias{date_to_isoquarter_n.character}
\alias{date_to_isoquarter_n.Date}
\title{ISO quarter (numeric) from Date object}
\usage{
date_to_isoquarter_n(x = lubridate::today())

\method{date_to_isoquarter_n}{default}(x)

\method{date_to_isoquarter_n}{character}(x = lubridate::today())

\method{date_to_isoquarter_n}{Date}(x = lubridate::today())
}
\arguments{
\item{x}{a Date object or string, in the form of 'yyyy-mm-dd'}
}
\value{
ISO quarter in numeric
}
\description{
ISO quarter (numeric) from Date object
}
\examples{
date_to_isoquarter_n("2021-08-11")
date_to_isoquarter_n(lubridate::today())
}
