% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_date_year_week.R
\name{date_to_isoyear_c}
\alias{date_to_isoyear_c}
\title{ISO year (character) from Date object}
\usage{
date_to_isoyear_c(x = lubridate::today())
}
\arguments{
\item{x}{a Date object or string, in the form of 'yyyy-mm-dd'}
}
\value{
ISO year in character
}
\description{
ISO year (character) from Date object
}
\examples{
date_to_isoyear_c('2021-08-11')
date_to_isoyear_c(lubridate::today())
}
