% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_date_year_week.R
\name{date_to_isoyearweek_c}
\alias{date_to_isoyearweek_c}
\title{ISO year and week (character) from Date object}
\usage{
date_to_isoyearweek_c(x = lubridate::today())
}
\arguments{
\item{x}{a Date object or string, in the form of 'yyyy-mm-dd'}
}
\value{
ISO year and week in character
}
\description{
ISO year and week (character) from Date object
}
\examples{
date_to_isoyearweek_c('2021-08-11')
date_to_isoyearweek_c(lubridate::today())
}
