% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_season.R
\name{date_to_season_c}
\alias{date_to_season_c}
\title{Date to season.}
\usage{
date_to_season_c(x)
}
\arguments{
\item{x}{date}
}
\description{
Date to season.
}
\examples{
date_to_season_c(c("2021-01-01","2021-12-01"))
}
