% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_season.R
\name{date_to_seasonweek_n}
\alias{date_to_seasonweek_n}
\title{Date to season week.}
\usage{
date_to_seasonweek_n(x)
}
\arguments{
\item{x}{date}
}
\description{
Date to season week.
}
\examples{
date_to_seasonweek_n(c("2021-01-01", "2021-12-01"))
}
