% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates_by_isoyearweek.R
\docType{data}
\name{dates_by_isoyearweek}
\alias{dates_by_isoyearweek}
\title{Dates of different days within isoyearweeks}
\format{
\describe{
\item{isoyear}{Isoyear.}
\item{isoyearweek}{Isoweek-isoyear.}
\item{mon}{Date of Monday.}
\item{tue}{Date of Tuesday.}
\item{wed}{Date of Wednesday.}
\item{thu}{Date of Thursday.}
\item{fri}{Date of Friday.}
\item{sat}{Date of Saturday.}
\item{sun}{Date of Sunday.}
\item{weekdays}{List of dates from Mon-Fri}
\item{weekend}{List of dates from Sat-Sun}
\item{days}{List of dates from Mon-Sun}
}
}
\usage{
dates_by_isoyearweek
}
\description{
Dates of different days within isoyearweeks
}
\examples{
# Constructing a vector of dates without removing the Date class
do.call("c", dates_by_isoyearweek[isoyearweek \%in\% c("2021-01", "2021-02")]$weekdays)
}
\keyword{datasets}
