% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoxx.R
\name{isoweek_c}
\alias{isoweek_c}
\title{ISO week (character) from Date object}
\usage{
isoweek_c(x = lubridate::today())
}
\arguments{
\item{x}{a Date object or string, in the form of 'yyyy-mm-dd'}
}
\value{
ISO week in character
}
\description{
ISO week (character) from Date object
}
\examples{
isoyear_c('2021-08-11')
isoyear_c(lubridate::today())
}
