% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoxx.R
\name{isoweek_n}
\alias{isoweek_n}
\title{ISO week (numeric) from Date object}
\usage{
isoweek_n(x = lubridate::today())
}
\arguments{
\item{x}{a Date object or string, in the form of 'yyyy-mm-dd'}
}
\value{
ISO week in numeric
}
\description{
ISO week (numeric) from Date object
}
\examples{
isoweek_n('2021-08-11')
isoweek_n(lubridate::today())
}
