% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_season.R
\name{isoweek_to_seasonweek_n}
\alias{isoweek_to_seasonweek_n}
\title{ISO week to season week (numeric). Season week 1 is natural week 30.}
\usage{
isoweek_to_seasonweek_n(isoweek)
}
\arguments{
\item{isoweek}{ISO week in a year, between 1 and 53. ISO week 53 is season week 23.5}
}
\description{
ISO week to season week (numeric). Season week 1 is natural week 30.
}
