% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoxx.R
\name{isoyear_n}
\alias{isoyear_n}
\title{ISO year (numeric) from Date object}
\usage{
isoyear_n(x = lubridate::today())
}
\arguments{
\item{x}{a Date object or string, in the form of 'yyyy-mm-dd'}
}
\value{
ISO year in numeric
}
\description{
ISO year (numeric) from Date object
}
\examples{
isoyear_n('2021-08-11')
isoyear_n(lubridate::today())
}
