% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_date_isoweek_isoyear.R
\name{isoyear_to_last_isoyearweek_c}
\alias{isoyear_to_last_isoyearweek_c}
\alias{isoyear_to_last_isoyearweek_c.default}
\alias{isoyear_to_last_isoyearweek_c.character}
\alias{isoyear_to_last_isoyearweek_c.numeric}
\title{Last ISO yearweek (character) in ISO year}
\usage{
isoyear_to_last_isoyearweek_c(x)

\method{isoyear_to_last_isoyearweek_c}{default}(x)

\method{isoyear_to_last_isoyearweek_c}{character}(x)

\method{isoyear_to_last_isoyearweek_c}{numeric}(x)
}
\arguments{
\item{x}{ISO year, e.g. 2020}
}
\value{
ISO year-week in character, of the last ISO year
}
\description{
Returns the last isoyearweek in the isoyear
}
\examples{
isoyear_to_last_isoyearweek_c(c(2019, 2019, 2020, 2021))
}
