% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_date_isoweek_isoyear.R
\name{isoyearweek_to_isoquarter_n}
\alias{isoyearweek_to_isoquarter_n}
\alias{isoyearweek_to_isoquarter_n.default}
\alias{isoyearweek_to_isoquarter_n.character}
\title{ISO yearweek to quarter (numeric)}
\usage{
isoyearweek_to_isoquarter_n(x)

\method{isoyearweek_to_isoquarter_n}{default}(x)

\method{isoyearweek_to_isoquarter_n}{character}(x)
}
\arguments{
\item{x}{Year-week, e.g. "2020-19" for 19th week in 2020}
}
\value{
ISO quarter in numeric
}
\description{
This function breaks the string connected with '-' into year/quarter
}
\examples{
isoyearweek_to_isoquarter_n('2020-19')
}
