% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_date_isoweek_isoyear.R
\name{isoyearweek_to_isoyear_c}
\alias{isoyearweek_to_isoyear_c}
\alias{isoyearweek_to_isoyear_c.default}
\alias{isoyearweek_to_isoyear_c.character}
\title{ISO yearweek to year (character)}
\usage{
isoyearweek_to_isoyear_c(x)

\method{isoyearweek_to_isoyear_c}{default}(x)

\method{isoyearweek_to_isoyear_c}{character}(x)
}
\arguments{
\item{x}{Year-week, e.g. "2020-19" for 19th week in 2020}
}
\value{
ISO year in character
}
\description{
This function breaks the string connected with '-' into year/week
}
\examples{
isoyearweek_to_isoyear_c('2020-10')
}
