% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_date_isoweek_isoyear.R
\name{isoyearweek_to_isoyearquarter_c}
\alias{isoyearweek_to_isoyearquarter_c}
\alias{isoyearweek_to_isoyearquarter_c.default}
\alias{isoyearweek_to_isoyearquarter_c.character}
\title{ISO yearweek to ISO yearquarter (character)}
\usage{
isoyearweek_to_isoyearquarter_c(x)

\method{isoyearweek_to_isoyearquarter_c}{default}(x)

\method{isoyearweek_to_isoyearquarter_c}{character}(x)
}
\arguments{
\item{x}{Year-week, e.g. "2020-19" for 19th week in 2020}
}
\value{
ISO yearquarter in character
}
\description{
ISO yearweek to ISO yearquarter (character)
}
\examples{
isoyearweek_to_isoyearquarter_c('2020-19')
}
