% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_date_isoweek_isoyear.R
\name{isoyearweek_to_last_date}
\alias{isoyearweek_to_last_date}
\title{Last date in ISO yearweek}
\usage{
isoyearweek_to_last_date(x)
}
\arguments{
\item{x}{ISO yearweek, e.g. "2020-19" for 19th week in 2020}
}
\description{
Returns the last date in the isoyearweek
}
\examples{
isoyearweek_to_last_date(c("2019-19", "2020-01"))
}
