% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearweek_split.R
\name{isoyearweek_to_week_c}
\alias{isoyearweek_to_week_c}
\title{isoyearweek_to_week_c
isoyearweek to week (character)
This function breaks the string connected with '-' into year/week}
\usage{
isoyearweek_to_week_c(yrwk)
}
\arguments{
\item{yrwk}{Year-week, e.g. "2020-19" for 19th week in 2020}
}
\description{
isoyearweek_to_week_c
isoyearweek to week (character)
This function breaks the string connected with '-' into year/week
}
