% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunday.R
\name{keep_sundays}
\alias{keep_sundays}
\title{keep_sundays}
\usage{
keep_sundays(dates, format = "Uke isoweek_c-1/isoweek_c", keep_delete = TRUE)
}
\arguments{
\item{dates}{dates (later)}

\item{format}{format (later)}

\item{keep_delete}{keep_delete (later)}
}
\description{
keep_sundays
}
