% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunday.R
\name{keep_sundays_and_latest_date}
\alias{keep_sundays_and_latest_date}
\title{keep_sundays_and_latest_date}
\usage{
keep_sundays_and_latest_date(
  dates,
  format = "Uke isoweek_c-1/isoweek_c",
  keep_delete = TRUE
)
}
\arguments{
\item{dates}{dates (later)}

\item{format}{format (later)}

\item{keep_delete}{keep_delete (later)}
}
\description{
keep_sundays_and_latest_date
}
