% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunday.R
\name{keep_sundays_and_latest_date_internal}
\alias{keep_sundays_and_latest_date_internal}
\title{keep_sundays_and_latest_date_internal}
\usage{
keep_sundays_and_latest_date_internal(
  dates,
  format = "Uke isoweek_c-1/isoweek_c",
  keep_delete = TRUE,
  keep_latest_date = TRUE
)
}
\arguments{
\item{dates}{dates (later)}

\item{format}{format (later)}

\item{keep_delete}{keep_delet (later)}

\item{keep_latest_date}{keep_latest_date (later)}
}
\description{
keep_sundays_and_latest_date_internal
}
