% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_date_year_week.R
\name{last_date_in_isoyearweek}
\alias{last_date_in_isoyearweek}
\title{Last date in ISO yearweek}
\usage{
last_date_in_isoyearweek(x)
}
\arguments{
\item{x}{ISO yearweek, e.g. "2020-19" for 19th week in 2020}
}
\description{
Returns the last date in the isoyearweek
}
\examples{
last_date_in_isoyearweek(c("2019-19", "2020-01"))
}
