% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_date_year_week.R
\name{last_isoyearweek_c_in_isoyear}
\alias{last_isoyearweek_c_in_isoyear}
\title{Last ISO yearweek (character) in ISO year}
\usage{
last_isoyearweek_c_in_isoyear(x)
}
\arguments{
\item{x}{ISO year, e.g. 2020}
}
\description{
Returns the last isoyearweek in the isoyear
}
\examples{
last_isoyearweek_c_in_isoyear(c(2019, 2019, 2020, 2021))
}
