% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_data_documentation.R
\docType{data}
\name{nor_workdays_by_isoyearweek}
\alias{nor_workdays_by_isoyearweek}
\title{Norwegian workdays and holidays by isoyearweek}
\format{
\describe{
\item{isoyearweek}{YYYY-WW}
\item{public_holiday}{The proportion of the days within the isoyearweek that are public holidays}
\item{freeday}{The proportion of the days within the isoyearweek that are either public holidays or Saturday/Sunday}
\item{workday}{1 minus freeday}
}
}
\usage{
nor_workdays_by_isoyearweek
}
\description{
Norwegian workdays and holidays by isoyearweek
}
\keyword{datasets}
