% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/now_c.R
\name{now_c}
\alias{now_c}
\title{Current time as character}
\usage{
now_c(format = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{format}{Defaults to \%Y-\%m-\%d \%H:\%M:\%S}
}
\description{
Current time as character
}
