% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_season.R
\name{seasonweek_to_isoweek_c}
\alias{seasonweek_to_isoweek_c}
\title{Season week to ISO week (character). Season week 1 is ISO week 30.}
\usage{
seasonweek_to_isoweek_c(seasonweek)
}
\arguments{
\item{seasonweek}{Season week in a year (numeric), between 1 and 52}
}
\value{
ISO week in character
}
\description{
Season week to ISO week (character). Season week 1 is ISO week 30.
}
\examples{
seasonweek_to_isoweek_c(31)
}
