% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{seasonweek_to_week_n}
\alias{seasonweek_to_week_n}
\title{seasonweek_to_week_n
Season week to natural week. Season week 1 is natural week 30.}
\usage{
seasonweek_to_week_n(seasonweek)
}
\arguments{
\item{seasonweek}{Season week in a year}
}
\description{
seasonweek_to_week_n
Season week to natural week. Season week 1 is natural week 30.
}
