% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{week_to_seasonweek_n}
\alias{week_to_seasonweek_n}
\title{week_to_seasonweek_n
Natural week to season week. Season week 1 is natural week 30.}
\usage{
week_to_seasonweek_n(week)
}
\arguments{
\item{week}{Natural week in a year}
}
\description{
week_to_seasonweek_n
Natural week to season week. Season week 1 is natural week 30.
}
