% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_dates_isoyearweek.R
\docType{data}
\name{world_dates_isoyearweek}
\alias{world_dates_isoyearweek}
\title{Dates of different days within isoweekyears}
\format{
\describe{
\item{isoyear}{Isoyear.}
\item{isoyearweek}{Isoweek-isoyear.}
\item{mon}{Date of Monday.}
\item{tue}{Date of Tuesday.}
\item{wed}{Date of Wednesday.}
\item{thu}{Date of Thursday.}
\item{fri}{Date of Friday.}
\item{sat}{Date of Saturday.}
\item{sun}{Date of Sunday.}
}
}
\usage{
world_dates_isoyearweek
}
\description{
Dates of different days within isoweekyears
}
\keyword{datasets}
