% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_date_calmonth_calyear.R
\name{date_to_calyearmonth_c}
\alias{date_to_calyearmonth_c}
\title{Date -> calyearmonth (character)}
\usage{
date_to_calyearmonth_c(x = lubridate::today())
}
\arguments{
\item{x}{a Date object or string, in the form of 'yyyy-mm-dd'}
}
\value{
calyearmonth ("YYYY-MXX")
}
\description{
Date -> calyearmonth (character)
}
\examples{
date_to_calyearmonth_c('2021-08-11')
date_to_calyearmonth_c(lubridate::today())
}
