% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_date_isoweek_isoyear.R
\name{isoyear_to_last_date}
\alias{isoyear_to_last_date}
\title{Last date in ISO year}
\usage{
isoyear_to_last_date(x)
}
\arguments{
\item{x}{ISO year, e.g. 2020}
}
\description{
Returns the last date in the isoyear
}
\examples{
isoyear_to_last_date(c(2019, 2019, 2020, 2021))
}
