% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_date_year_week.R
\name{isoyear_to_last_isoyearweek_c}
\alias{isoyear_to_last_isoyearweek_c}
\title{Last ISO yearweek (character) in ISO year}
\usage{
isoyear_to_last_isoyearweek_c(x)
}
\arguments{
\item{x}{ISO year, e.g. 2020}
}
\description{
Returns the last isoyearweek in the isoyear
}
\examples{
isoyear_to_last_isoyearweek_c(c(2019, 2019, 2020, 2021))
}
