% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays.R
\docType{data}
\name{norway_workdays_by_isoyearweek}
\alias{norway_workdays_by_isoyearweek}
\title{Norwegian workdays and holidays by isoyearweek}
\format{
\describe{
\item{isoyearweek}{YYYY-WW}
\item{mon_to_fri}{The proportion of the days within the isoyearweek that are Monday to Friday (5/7)}
\item{sat_to_sun}{The proportion of the days within the isoyearweek that are Saturday to Sunday (2/7)}
\item{public_holiday}{The proportion of the days within the isoyearweek that are public holidays}
\item{freeday}{The proportion of the days within the isoyearweek that are either public holidays or Saturday/Sunday}
\item{workday}{1 minus freeday}
}
}
\usage{
norway_workdays_by_isoyearweek
}
\description{
Norwegian workdays and holidays by isoyearweek
}
\keyword{datasets}
